Added support for creating the initial Projects and Activities in a project based on the new Item.Presets.xml definition. No longer are all new databases populated with a hardcoded Default Activity and Default Project. One of any number of several presets will be created depending on either: 1) the use of the New Database wizard or 2) the user bypassing the wizard and getting the default preset.
Other notes:
- I'm second-guessing the ubiquitousness of LocationId.
- I'm also starting to put Projects first when talking about the items.
  That's the last remaining psychological remnant of 'Tasks'.

-----------------------------------------------------------------------
TESTING
- WIP
-----------------------------------------------------------------------
DETAILS

- Classes/Datatypes.cs (edit)
  * Changed the FileCreate and FileUpdate structs to classes.
  * Created a base class (FileBaseOptions) to hold the common valuees.

- Classes/File.cs (edit)
  * Aded an options argument to PopulateLocation().
  * Removed 'populate' from CreateTable("Activity") and ("Project").
  * Added PopulateItems(). This method takes the user-selected preset
    value (or the default preset of '1') and opens the Item.Presets.xml
    file to create the initial Projects and Activities.

- Classes/File.Upgrade.cs (edit)
  * Changed Options to UpgradeOptions to not conflict with CreateOptions.

- Classes/Item.cs (edit)
  * Changed public properties with proper accessors.
  * Load-consturctor handles non-existent Items passed to it.
  * Updated Create() for Schema 3.0.0.1.
  * SetSecondsElapsedToday doesn't bother to do anything if Journal 
    doesn't (yet) exist.

- Classes/Timekeeper.cs (edit)
  * Added an experimental DoubleWarn() to write the same message to the
    log file as well as ShowMessage. Not sold on it.
  * Added CurrentTimeZoneId() to return you know what.

- Classes/Widgets.cs (edit)
  * item.Create() now returns a long.

- Forms/Main.Action.cs (edit)
  * Using Timekeeper.DoubleWarn() in CheckDatabase.

- Forms/Main.cs (edit)
  * XML test code. Will remove later.

- Forms/Main.Designer.cs (edit)
- Forms/Main.resx (edit)
  * XML test code menu item. Will remove later.

- Forms/Main.Dialog.cs (edit)
  * Dialog_NewFile now sets up full CreateOptions when the wizard is not
    being used (e.g., Location and Presets are now required).
  * Added Application.DoEvents() to paint the main window before heading
    into the potentially lengthy CreateFile.

- Forms/NewWizard.cs (edit)
  * Removed FileName as its own property since it's redundant with 
    CreateOptions.FileName.

- Properties/Resources.Designer.cs (edit)
- Properties/Resources.resx (edit)
  * Added Item.Presets.xml as a file resource.

- Resources/SQL/3.0/Activity.Insert.sql (edit)
- Resources/SQL/3.0/Project.Insert.sql (edit)
  * No longer used. Will remove later.

- Resources/SQL/3.0/Journal.Create.sql (edit)
  * The transient 'original' time stamps are now nullable.

- Resources/SQL/3.0/Meta.Insert.sql (edit)
  * Changed base Upgraded date/time to UTC.

- Resources/XML/Item.Presets.xml (add)
  * First stab at this file. Fully functional in concept, but needs to
    be fleshed out to support the presets.

- timekeeper.csproj (edit)
  * Added System.Xml.Linq, but not using it.

-----------------------------------------------------------------------
Project...: Timekeeepr
Components: New Database Wizard
Target....: 3.0

Tickets Addressed:
  - 1275: New Database Wizard

[git-p4: depot-paths = "//depot/timekeeper/": change = 300]

+++ timekeeper/Classes/Widgets.cs
@@ -123,7 +123,7 @@ namespace Timekeeper.Classes
 
             int Result = TREES_ITEM_CREATED;
 
-            int CreateStatus = item.Create();
+            long CreateStatus = item.Create();
             switch (CreateStatus) {
                 case 1:
                     TreeNode foo = AddItemToTree(tree, ParentNode, item, imageIndex);
